+++ src/main/java/pl/plajer/villagedefense/commands/arguments/game/LeaderboardArgument.java
@@ -37,6 +37,7 @@ import pl.plajer.villagedefense.commands.arguments.ArgumentsRegistry;
 import pl.plajer.villagedefense.commands.arguments.data.CommandArgument;
 import pl.plajer.villagedefense.commands.completion.CompletableArgument;
 import pl.plajer.villagedefense.handlers.language.Messages;
+import pl.plajer.villagedefense.user.data.MysqlManager;
 
 /**
  * @author Plajer
@@ -94,7 +95,7 @@ public class LeaderboardArgument {
         if (registry.getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.DATABASE_ENABLED)) {
           try (Connection connection = registry.getPlugin().getMysqlDatabase().getConnection();
                Statement statement = connection.createStatement();
-               ResultSet set = statement.executeQuery("SELECT name FROM playerstats WHERE UUID='" + current.toString() + "'")) {
+               ResultSet set = statement.executeQuery("SELECT name FROM "+ ((MysqlManager) registry.getPlugin().getUserManager().getDatabase()).getTableName() +" WHERE UUID='" + current.toString() + "'")) {
             if (set.next()) {
               sender.sendMessage(formatMessage(statistic, set.getString(1), i + 1, stats.get(current)));
               continue;