+++ src/main/java/pl/plajer/villagedefense/user/data/MysqlManager.java
@@ -28,11 +28,13 @@ import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
 
+import org.bukkit.configuration.file.FileConfiguration;
 import pl.plajer.villagedefense.Main;
 import pl.plajer.villagedefense.api.StatsStorage;
 import pl.plajer.villagedefense.user.User;
 import pl.plajer.villagedefense.utils.MessageUtils;
 import pl.plajerlair.commonsbox.database.MysqlDatabase;
+import pl.plajerlair.commonsbox.minecraft.configuration.ConfigUtils;
 
 /**
  * @author Plajer
@@ -50,7 +52,7 @@ public class MysqlManager implements UserDatabase {
     Bukkit.getScheduler().runTaskAsynchronously(plugin, () -> {
       try (Connection connection = database.getConnection();
            Statement statement = connection.createStatement()) {
-        statement.executeUpdate("CREATE TABLE IF NOT EXISTS `playerstats` (\n"
+        statement.executeUpdate("CREATE TABLE IF NOT EXISTS `"+getTableName()+"` (\n"
             + "  `UUID` char(36) NOT NULL PRIMARY KEY,\n"
             + "  `name` varchar(32) NOT NULL,\n"
             + "  `kills` int(11) NOT NULL DEFAULT '0',\n"
@@ -64,7 +66,7 @@ public class MysqlManager implements UserDatabase {
 
         //temporary workaround
         try {
-          statement.executeUpdate("ALTER TABLE playerstats ADD `name` text NOT NULL");
+          statement.executeUpdate("ALTER TABLE "+getTableName()+" ADD `name` text NOT NULL");
         } catch (MySQLSyntaxErrorException e) {
           if (!e.getMessage().contains("Duplicate column name")) {
             plugin.getLogger().log(Level.WARNING, "Could not connect to MySQL database! Cause: {0} ({1})", new Object[] {e.getSQLState(), e.getErrorCode()});
@@ -82,7 +84,7 @@ public class MysqlManager implements UserDatabase {
   @Override
   public void saveStatistic(User user, StatsStorage.StatisticType stat) {
     Bukkit.getScheduler().runTaskAsynchronously(plugin, () ->
-        database.executeUpdate("UPDATE playerstats SET " + stat.getName() + "=" + user.getStat(stat) + " WHERE UUID='" + user.getPlayer().getUniqueId().toString() + "';"));
+        database.executeUpdate("UPDATE "+getTableName()+" " + stat.getName() + "=" + user.getStat(stat) + " WHERE UUID='" + user.getPlayer().getUniqueId().toString() + "';"));
   }
 
   @Override
@@ -91,7 +93,7 @@ public class MysqlManager implements UserDatabase {
       String uuid = user.getPlayer().getUniqueId().toString();
       try (Connection connection = database.getConnection()) {
         Statement statement = connection.createStatement();
-        ResultSet rs = statement.executeQuery("SELECT * from playerstats WHERE UUID='" + uuid + "'");
+        ResultSet rs = statement.executeQuery("SELECT * from "+getTableName()+" WHERE UUID='" + uuid + "'");
         if (rs.next()) {
           //player already exists - get the stats
           for (StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
@@ -103,7 +105,7 @@ public class MysqlManager implements UserDatabase {
           }
         } else {
           //player doesn't exist - make a new record
-          statement.executeUpdate("INSERT INTO playerstats (UUID,name) VALUES ('" + uuid + "','" + user.getPlayer().getName() + "')");
+          statement.executeUpdate("INSERT INTO "+getTableName()+" (UUID,name) VALUES ('" + uuid + "','" + user.getPlayer().getName() + "')");
           for (StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
             if (!stat.isPersistent()) {
               continue;
@@ -115,6 +117,11 @@ public class MysqlManager implements UserDatabase {
         plugin.getLogger().log(Level.WARNING, "Could not connect to MySQL database! Cause: {0} ({1})", new Object[] {e.getSQLState(), e.getErrorCode()});
       }
     });
+  }
+
+  public String getTableName() {
+    FileConfiguration config = ConfigUtils.getConfig(plugin, "mysql");
+    return config.getString("table", "playerstats");
   }